src/requirements.txt
alembic==1.7.1
anyio==3.6.1
asgiref==3.5.2
asyncpg
astunparse==1.6.3
attrs==21.4.0
bleach==4.1.0
cachetools==4.2.4
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography
cycler==0.11.0
distlib==0.3.4
easydict==1.9
environ-config==21.2.0
fastapi==0.78.0
filelock==3.7.1
flatbuffers==2.0
fonttools==4.33.3
gast==0.4.0
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
identify==2.5.1
idna==3.3
importlib-metadata==4.11.4
kiwisolver==1.4.3
libclang==14.0.1
lxml==4.6.3
markdown==3.3.7
nodeenv==1.6.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
platformdirs==2.5.2
pre-commit==2.19.0
proto-plus==1.20.3;
protobuf==3.19.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pyyaml==6.0
psycopg2-binary==2.9.3
requests-oauthlib==1.3.1
requests==2.28.0
setuptools-scm==7.0.2
six==1.16.0
starlette==0.19.1
SQLAlchemy==1.4.15
sqlalchemy_utils
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
uvicorn==0.17.6

zipp==3.8.0

fastapi-pagination
sqladmin
sqlalchemy-stubs
mypy-protobuf
mypy==0.790
flake8==3.8.4
isort==5.6.4
pytest-isort==1.2.0
pytest
pytest-cov
pytest-mock
pytest-asyncio
responses
safety
ipython
coverage